requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
firebase-admin==5.2.0
Werkzeug==2.0.3
itsdangerous==2.0.1
PyMySQL==1.0.2
